Share the Tidal provider mock between test modules
