+++ tests/providers/tidal/conftest.py
@@ -1,18 +1,75 @@
 """Fixtures for Tidal provider tests."""
 
-from collections.abc import Generator
-from unittest.mock import AsyncMock, Mock, patch
+from __future__ import annotations
+
+from collections.abc import AsyncGenerator, Generator
+from typing import TYPE_CHECKING, Any
+from unittest.mock import AsyncMock, MagicMock, Mock, patch
 
 import pytest
+from music_assistant_models.media_items import ItemMapping
 
 from music_assistant.providers.tidal.media import TidalMediaManager
+
+if TYPE_CHECKING:
+    from music_assistant_models.enums import MediaType
+
+
+@pytest.fixture
+def provider_mock() -> Mock:
+    """Return a mock Tidal provider with an authenticated session and wired collaborators."""
+    provider = Mock()
+    provider.domain = "tidal"
+    provider.instance_id = "tidal_instance"
+
+    provider.auth.user_id = "12345"
+    provider.auth.country_code = "US"
+    provider.auth.access_token = "token"
+    provider.auth.session_id = "session"
+    provider.auth.user.profile_name = "Test User"
+    provider.auth.user.user_name = "Test User"
+    provider.auth.ensure_valid_token = AsyncMock(return_value=True)
+    provider.auth.refresh_token = AsyncMock()
+
+    provider.api = AsyncMock()
+    provider.api.get.return_value = {}
+    # paginate is an async generator, so it cannot be an AsyncMock child: it yields the
+    # items a test assigns to paginate.return_value
+    provider.api.paginate = MagicMock()
+
+    async def paginate(*_args: Any, **_kwargs: Any) -> AsyncGenerator[Any]:
+        for item in provider.api.paginate.return_value:
+            yield item
+
+    provider.api.paginate.side_effect = paginate
+    provider.api.paginate.return_value = []
+
+    provider.get_track = AsyncMock()
+
+    def get_item_mapping(media_type: MediaType, key: str, name: str) -> ItemMapping:
+        return ItemMapping(
+            media_type=media_type,
+            item_id=key,
+            provider=provider.instance_id,
+            name=name,
+        )
+
+    provider.get_item_mapping.side_effect = get_item_mapping
+
+    provider.mass.http_session = AsyncMock()
+    provider.mass.metadata.locale = "en_US"
+    provider.mass.config.get_provider_configs = AsyncMock(return_value=[])
+    provider.mass.cache.get = AsyncMock(return_value=None)
+    provider.mass.cache.set = AsyncMock()
+    provider.mass.cache.delete = AsyncMock()
+    provider.mass.music.tracks.get_library_item_by_prov_id = AsyncMock(return_value=None)
+
+    return provider
 
 
 @pytest.fixture
 def media_manager(provider_mock: Mock) -> TidalMediaManager:
     """Return a TidalMediaManager instance."""
-    # provider_mock is defined per test module and differs between them, so a module
-    # requesting this fixture must bring its own
     return TidalMediaManager(provider_mock)
