+++ tests/providers/tidal/test_media.py
@@ -1,48 +1,11 @@
 """Test Tidal Media Manager."""
 
-from typing import Any
-from unittest.mock import AsyncMock, MagicMock, Mock, patch
+from unittest.mock import Mock, patch
 
-import pytest
 from music_assistant_models.enums import MediaType
 from music_assistant_models.errors import RetriesExhausted
-from music_assistant_models.media_items import ItemMapping
 
 from music_assistant.providers.tidal.media import TidalMediaManager
-
-
-@pytest.fixture
-def provider_mock() -> Mock:
-    """Return a mock provider."""
-    provider = Mock()
-    provider.domain = "tidal"
-    provider.instance_id = "tidal_instance"
-    provider.auth.user_id = "12345"
-    provider.auth.country_code = "US"
-    provider.api = AsyncMock()
-    provider.api.get.return_value = {}
-    provider.api.paginate = MagicMock()
-
-    async def async_iter(*_args: Any, **_kwargs: Any) -> Any:
-        for item in provider.api.paginate.return_value:
-            yield item
-
-    provider.api.paginate.side_effect = async_iter
-    provider.api.paginate.return_value = []
-
-    provider.logger = Mock()
-
-    def get_item_mapping(media_type: MediaType, key: str, name: str) -> ItemMapping:
-        return ItemMapping(
-            media_type=media_type,
-            item_id=key,
-            provider=provider.instance_id,
-            name=name,
-        )
-
-    provider.get_item_mapping.side_effect = get_item_mapping
-
-    return provider
 
 
 @patch("music_assistant.providers.tidal.media.parse_artist")
